feat: Decimal

+++ exc-binance/src/futures_api/http/trading.rs
@@ -1,6 +1,7 @@
 use super::super::types::{OrderType, TimeInForce};
 use exc_util::interface::{ApiKind, Method, Rest};
 use exc_util::types::order::{OrderSide, OrderStatus};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, DisplayFromStr};
 
@@ -13,10 +14,8 @@ pub struct PlaceOrderRequest {
     pub side: OrderSide,
     pub r#type: OrderType,
     pub time_in_force: TimeInForce,
-    #[serde_as(as = "DisplayFromStr")]
-    pub quantity: f64,
-    #[serde_as(as = "DisplayFromStr")]
-    pub price: f64,
+    pub quantity: Decimal,
+    pub price: Decimal,
 }
 
 #[derive(Debug, Deserialize)]
