+++ exc-okx/src/api/http/trading.rs
@@ -1,5 +1,6 @@
 use super::super::types::{FuturesOpenType, OrderSide, OrderStatus, OrderType};
 use exc_util::interface::{ApiKind, Method, Rest};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, DisplayFromStr};
 
@@ -15,10 +16,8 @@ pub struct PlaceOrderRequest {
     pub side: OrderSide,
     /// market：市价单 limit：限价单 post_only：只做maker单 fok：全部成交或立即取消 ioc：立即成交并取消剩余 optimal_limit_ioc：市价委托立即成交并取消剩余（仅适用交割、永续）
     pub ord_type: OrderType,
-    #[serde_as(as = "DisplayFromStr")]
-    pub sz: f64,
-    #[serde_as(as = "DisplayFromStr")]
-    pub px: f64,
+    pub sz: Decimal,
+    pub px: Decimal,
     pub cl_ord_id: Option<String>,
 }
